requirements.txt
google-generativeai
streamlit
pillow
streamlit-option-menu
python-dotenv